# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Warehouse Management System API",
        "title": "Warehouse API",
        "contact": {},
        "version": "1.2"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/reports/store-product-quantities": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns total product quantity per store, with optional filters by product and category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Store product quantities report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "total_quantity",
                            "product_id",
                            "store_id"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for next page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StoreProductQuantityListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/reports/store-product-quantities/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Export filtered store product quantities report as an Excel (.xlsx) file. Supports the same query filters as the store product quantities report list endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Export store product quantities report to Excel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "total_quantity",
                            "product_id",
                            "store_id"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/reports/threshold-proximity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns products whose current stock quantity is at or below the warning threshold, including category, unit, and store count.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Threshold proximity report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "current_quantity",
                            "product_id"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for next page",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ThresholdProximityListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/reports/threshold-proximity/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "Bearer": []
                    }
                ],
                "description": "Export filtered threshold proximity report as an Excel (.xlsx) file. Supports the same query filters as the threshold proximity report list endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Export threshold proximity report to Excel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "current_quantity",
                            "product_id"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort direction",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "dto.StoreProductQuantityListResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StoreProductQuantityReport"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "next_cursor": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "dto.ThresholdProximityListResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThresholdProximityReport"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "next_cursor": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "models.StoreProductQuantityReport": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 2
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "store_id": {
                    "type": "integer",
                    "example": 3
                },
                "total_quantity": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "models.ThresholdProximityReport": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 2
                },
                "category_name": {
                    "type": "string",
                    "example": "Category A"
                },
                "current_quantity": {
                    "type": "integer",
                    "example": 5
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "product_name": {
                    "type": "string",
                    "example": "Product A"
                },
                "stores_count": {
                    "type": "integer",
                    "example": 2
                },
                "unit_id": {
                    "type": "integer",
                    "example": 3
                },
                "unit_name": {
                    "type": "string",
                    "example": "kg"
                },
                "warning_threshold": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API Key here",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "Bearer": {
            "description": "Enter JWT with Bearer prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}